pyproject.toml
[project]
name = "ups-mcp"
version = "1.1.0"
description = "MCP server for UPS APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.26.0",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.4",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/UPS-API/ups-mcp"

[project.scripts]
ups-mcp = "ups_mcp.server:main"
